Split check/mod.rs's module-inference and decl-checking methods into mod_p3.rs
